chore: remove VCS login related APIs
